[frontend] pr comments and refacto
